feat(contracts): Update contracts used, rinkeby addresses and config store
