+++ src/config/rinkeby.json
@@ -1,12 +1,9 @@
 {
-  "avatar": "0xE796D5650b664db3C58f778A7e19e5EDe794bf04",
-  "controller": "0xEB6589A274cB723D0354cEc6E72882a55a632499",
-  "reputation": "0x7774bBbf570De31F98673534aEd686E3Dd09B5Ff",
-  "votingMachine": "0x03708EFA0F32f506cF93B722C3eFDf907f3E63Bd",
-  "votingMachineToken": "0xa700BdAba48A3D96219247111B0b708Dc0b51033",
-  "masterWalletScheme": "0x200188373f1346c615089b8a880E64de2C0854eC",
-  "quickWalletScheme": "0x7cF7Ea17855EeD397A76548D90886d165737C2D8",
-  "permissionRegistry": "0x80b5aCD6d3Bb4386908fAa198b7317dC6233BAEe",
+  "avatar": "0xb4F966467C67F5A2676f8FafBc4E9A0940b8e87B",
+  "controller": "0x6076adBa18948b1cb37004a9C6Caec4Efda622D1",
+  "reputation": "0xA683750a9FCCAdcaA63367623B014321073B0155",
+  "permissionRegistry": "0xe3c2241c5bAA2F421EDBBc0734B098597BBF1439",
+  "votingMachine": "0x2467c2f5E07A122EF65fAEfb78B2ad1D3f3aCDC8",
   "multicall": "0xa1b0a36c7aE04A84EBF493e03dF0aC295eE90747",
-  "fromBlock": 8389437
+  "fromBlock": 8452678
 }
