+++ src/stores/ConfigStore.ts
@@ -1,8 +1,8 @@
 import { action, observable } from 'mobx';
 import RootStore from 'stores';
-import { getConfig } from '../config/contracts';
+import { getConfig } from '../config';
 import { _ } from 'lodash';
-import { CHAIN_NAME_BY_ID, DEFAULT_ETH_CHAIN_ID } from '../provider/connectors';
+import { CHAIN_NAME_BY_ID } from '../provider/connectors';
 
 export default class ConfigStore {
     @observable darkMode: boolean;
@@ -15,7 +15,7 @@ export default class ConfigStore {
     
     getActiveChainName() {
       const activeWeb3 = this.rootStore.providerStore.getActiveWeb3React();
-      return CHAIN_NAME_BY_ID[(activeWeb3 && activeWeb3.chainId) ? activeWeb3.chainId : DEFAULT_ETH_CHAIN_ID];
+      return activeWeb3 ? CHAIN_NAME_BY_ID[activeWeb3.chainId] : 'none';
     }
     
     getApiKeys() {
@@ -56,27 +56,12 @@ export default class ConfigStore {
     getVotingMachineAddress() {
       return getConfig(this.getActiveChainName()).votingMachine || "0x0000000000000000000000000000000000000000";
     }
-    getVotingMachineTokenAddress() {
-      return getConfig(this.getActiveChainName()).votingMachineToken || "0x0000000000000000000000000000000000000000";
+    getPermissionRegistryAddress() {
+      return getConfig(this.getActiveChainName()).permissionRegistry || "0x0000000000000000000000000000000000000000";
     }
     getMulticallAddress() {
       return getConfig(this.getActiveChainName()).multicall || "0x0000000000000000000000000000000000000000";
     }
-    getSchemeAddress(schemeName) {
-      return getConfig(this.getActiveChainName()).schemes[schemeName] || "0x0000000000000000000000000000000000000000";
-    }
-    
-    getSchemeName(schemeAddress) {
-      function swap(obj){
-        var ret = {};
-        for(var key in obj)
-          ret[obj[key]] = key;
-        return ret;
-      }
-      const schemeName = swap(getConfig(this.getActiveChainName()).schemes)[schemeAddress];
-      return (schemeName.charAt(0).toUpperCase() + schemeName.slice(1)).replace(/([A-Z])/g, ' $1').trim()
-    }
-    
     getStartBlock() {
       return getConfig(this.getActiveChainName()).fromBlock;
     }
